devops(clang-tidy): Adjust config to analyze header files correctly
It is common to only pass in source files to `clang-tidy`.
To also show diagnostics from header files, `-header-filter=.*` can be used.
However, if there are header files that don't get included anywhere, they would not get analyzed.
Imo this is a problem when a header-only utility gets added by one person and only later it will be used by another person - only when it gets used there will be warnings, but the original author should have fixed them in the first place.