devops(clang-tidy): Remove build-related flags from config
I thought it would be better to not have a strong dependency between the build system and `clang-tidy`.
But apparently it is more common that all warning and build flags are just handled by CMake, and `clang-tidy` should then use those via the `compile_commands.json`.
I also read that `clang-tidy` is designed to understand very different flags, so even if other compilers were used for building, it should be fine.