devops(clang-tidy): Restore `va_arg` warning
We should only completely disable warnings if we disagree with a warning in all cases.
We probably don't want to write variadic functions.

When we do need to call a variadic function from the C API, putting a `NOLINT` can be very helpful as a reminder that the compiler will not do any type checking here.

Also fix an alias having been accidentally moved to the "removed" section.