style(libftpp): Rename `boolean_test()` to `operator_bool()`
Because the function is in the public API of a derived class, naming it
like the operator discourages calling this function directly and leaves
it up to the derived class if only the implicit conversion should be
used.
Also makes it easier to understand for someone not familiar with the
safe bool idiom why this function exists.