perf: Avoid extra copy of extracted string
Copy elision is not possible when the return type still needs to be constructed; certainly not with the current implementation of `ft::expected`.

Optimally, the `std::string` could be passed arguments at construction of the `ft::excpected`.
But that's not possible atm because `ft::in_place_t` is not implemented yet.