test(writeBody): Fix failure trace by using SCOPED_TRACE
The benefit of this solution is that the failure trace shows both the line in the test helper function and the line in the main test function.
With the macros, it only showed the line of the main test function, so it was not clear which exact line in the test helper function failed.

https://google.github.io/googletest/advanced.html#adding-traces-to-assertions